refactor(sso): move SSO into a dedicated polar/auth/sso package
